multicluster-demo/k8s_troubleshooting_agent/pyproject.toml
[project]
name = "k8s-troubleshooting-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk[a2a]>=1.28.0",
    "uvicorn>=0.42.0",
]
